changed search as component.
